Reland add tentative WPT tests for stale while revalidate handling.
Add test to ensure that handling the fetch doesn't trigger stale
while revalidate loading.

Add test to ensure that scripts and css loaded trigger a stale
while revalidate cache hit.

The PR for the spec changes is here:
https://github.com/whatwg/fetch/pull/853

This reland removes the resource timing and moves to css (instead of
images) which were both unreliable.

BUG=348877

Change-Id: Ibabd8d3fd0295bedc8259594fc926da6ab5cfc43