Worker: Fix wpt/workers/semantics/run-a-worker/003.html for shared workers
Before this CL, the test expects that a 404 failure on shared worker script
fetch doesn't fire an error event. However, this doesn't obey the HTML spec
(see below), and actually Chrome and Firefox fail the test [1]. Note that
Edge and Safari haven't been implemented shared workers.

This CL changes the test to expect that the failure fires an error event.

<HTML spec from [2]>
12. Obtain script by switching on the value of options's type member:
  - "classic": Fetch a classic worker script given url, outside settings,
    destination, and inside settings.

  (...snip...)

  If the algorithm asynchronously completes with null, then:
    - Queue a task to fire an event named error at worker.
    - Run the environment discarding steps for inside settings.
    - Return.
</HTML spec>

[1] https://wpt.fyi/results/workers/semantics/run-a-worker/003.html?run_id=5697959937179648&run_id=5733815918002176&run_id=5135009983758336&run_id=5639735062036480
[2] https://html.spec.whatwg.org/multipage/workers.html#run-a-worker

Bug: 655458, 924041
Change-Id: Ia4dc694e5bff5daed9a80a963e6316293c86c8bc